deleting coded

+++ src/main/java/model/FileUpdate.java
@@ -9,11 +9,11 @@ public class FileUpdate {
 
     public static void saveUser(User user) throws IOException {
         String path = DB + user.getUsername();
+        File dir = new File(path);
 
-        //make other files if user is new
-        if(new File(path).mkdir()) {
-            new File(path + "/Inbox.txt").createNewFile();
-            new File(path + "/Sent.txt").createNewFile();
+        //make user directory if user is new
+        if(!dir.exists()) {
+            dir.mkdir();
         }
 
         //serialize updated user into file
@@ -25,19 +25,58 @@ public class FileUpdate {
 
     public static List<Conversation> getMail(ServerMessage message) throws IOException, ClassNotFoundException {
         String path = DB + message.getSender().getUsername() + "/" + message.getMessageType() + ".txt";
+        if (!(new File(path).exists()))
+            return null;
         ObjectInputStream ois = new ObjectInputStream(new FileInputStream(path));
         List<Conversation> list = (List<Conversation>) ois.readObject();
         ois.close();
         return list;
     }
 
+    public static void deleteConversation(File file, Conversation conversation) throws IOException, ClassNotFoundException {
+        if (file.exists()) {
+            ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(file, true));
+            ObjectInputStream ois = new ObjectInputStream(new FileInputStream(file));
+            List<Conversation> conversations = (List<Conversation>) ois.readObject();
+            ois.close();
+            conversations.remove(conversation);
+            oos.writeObject(conversations);
+            oos.flush();
+            oos.close();
+        }
+    }
+
+    public static void deleteMessage(File file, Conversation conversation, Email email) throws IOException, ClassNotFoundException {
+        if (file.exists()) {
+            ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(file, true));
+            ObjectInputStream ois = new ObjectInputStream(new FileInputStream(file));
+            List<Conversation> conversations = (List<Conversation>) ois.readObject();
+            ois.close();
+            for (int i = 0; i < conversations.size(); i++) {
+                if (conversations.get(i).equals(conversation)) {
+                    conversations.get(i).getMessages().remove(email);
+                    break;
+                }
+            }
+            oos.writeObject(conversations);
+            oos.flush();
+            oos.close();
+        }
+    }
+
     public static void updateMail(File file, Conversation conversation) throws IOException, ClassNotFoundException {
-        ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(file, true));
-        ObjectInputStream ois = new ObjectInputStream(new FileInputStream(file));
-        Object listObject = ois.readObject();
         List<Conversation> conversations = new ArrayList<>();
-        if (listObject != null) {
-            conversations = (List<Conversation>) listObject;
+        ObjectOutputStream oos;
+        ObjectInputStream ois;
+        if (!file.exists()) {
+            oos = new ObjectOutputStream(new FileOutputStream(file));
+            conversations.add(conversation);
+        }
+        else {
+            oos = new ObjectOutputStream(new FileOutputStream(file, true));
+            ois = new ObjectInputStream(new FileInputStream(file));
+            conversations = (List<Conversation>) ois.readObject();
+            ois.close();
             for (int i = 0; i < conversations.size(); i++) {
                 if (conversation.equals(conversations.get(i))) {
                     conversations.set(i, conversation);
@@ -47,12 +86,8 @@ public class FileUpdate {
                     conversations.add(conversation);
             }
         }
-        else {
-            conversations.add(conversation);
-        }
         oos.writeObject(conversations);
         oos.flush();
-        ois.close();
         oos.close();
     }
 }
